packages/dataraum-api/pyproject.toml
[project]
name = "dataraum"
version = "0.1.0"
description = "Rich metadata context engine for AI-driven data analytics"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "Apache-2.0" }
authors = [
    { name = "DataRaum", email = "hello@dataraum.com" }
]
keywords = ["data", "metadata", "context", "ai", "analytics", "duckdb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Data processing
    "duckdb>=1.0.0",
    "pyarrow>=15.0.0",
    "pint>=0.23",
    "pandas>=2.0.0",
    # Database (SQLAlchemy + SQLite backend)
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    # API
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # MCP
    "mcp>=0.1.0",
    # Configuration
    "pyyaml>=6.0.0",
    # TDA (Topological Data Analysis)
    "scipy>=1.11.0",
    "ripser>=0.6.0",
    "persim>=0.3.0",
    # Time Series Analysis
    "statsmodels>=0.14.0",
    "ruptures>=1.1.0",
    # Utilities
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "networkx>=3.6",
    "anthropic>=0.76.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.98.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

# LLM providers (install what you need)
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.50.0"]
llm = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
]
postgres = ["asyncpg>=0.29.0"]

# Statistical quality metrics (ML-based anomaly detection)
statistical-quality = [
    "scikit-learn>=1.3.0",
]

# Synthetic data generation for privacy (SDV) - heavy dependency with PyTorch
privacy = ["sdv>=1.17.0"]

# Common optional features (excluding heavy privacy/SDV dependencies)
complete = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "asyncpg>=0.29.0",
    "scikit-learn>=1.3.0",
]

# Everything including heavy ML dependencies
all = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "sdv>=1.17.0",
    "asyncpg>=0.29.0",
    "scikit-learn>=1.3.0",
]

[project.scripts]
dataraum = "dataraum.cli:app"

[project.urls]
Homepage = "https://github.com/dataraum/dataraum-context"
Documentation = "https://docs.dataraum.com/context"
Repository = "https://github.com/dataraum/dataraum-context"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.98.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "pandas-stubs>=2.3.2.250926",
    "types-pyyaml>=6.0.12.20250915",
    "scipy-stubs>=1.16.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dataraum"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (may be slower)",
    "slow: marks tests as slow (use real data, skip in CI)",
]

[tool.ruff]
target-version = "py314"
line-length = 100
extend-exclude = [
    "../../prototypes/",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["dataraum"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_ignores = false  # Allow unused ignores during migration
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

# Ignore attr-defined errors for re-exported legacy types
# These are false positives due to exec()-based re-export in core.models.__init__
[[tool.mypy.overrides]]
module = "dataraum.core.models"

[[tool.mypy.overrides]]
module = [
    "duckdb.*",
    "mcp.*",
    "anthropic.*",
    "openai.*",
    "sdv.*",
    "hamilton.*",
    "ripser.*",
    "persim.*",
    "scipy.*",
    "sklearn.*",
    "statsmodels.*",
    "ruptures.*",
    "pyarrow.*",
    "networkx.*",
]
ignore_missing_imports = true

# Unfinished features that depend on not-yet-implemented context module
[[tool.mypy.overrides]]
module = [
    "dataraum.llm.features.queries",
    "dataraum.llm.features.summary",
]
ignore_errors = true

# Test files don't need strict type checking
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
ignore_missing_imports = true
